+++ data-migrations/2026-06-01-1542-remove-logoURI-from-affiliations.sql
@@ -1,0 +1,3 @@
+-- Remove the logoURI column from the affiliation table, as it is now constructed
+-- from the CDN URL and logoName (s3 key) within the resolver.
+ALTER TABLE `affiliations` DROP COLUMN `logoURI`;